Move HpgRefProduct class to its own files.
